Refactor CLI structure: consolidate command handling, improve argument parsing, and enhance contract management
