+++ odra-cli/src/entry_point/cmd_args.rs
@@ -1,8 +1,8 @@
 use odra::schema::casper_contract_schema::Entrypoint;
 
 use crate::cmd::args::CommandArg;
+use crate::custom_types::CustomTypeSet;
 use crate::entry_point::utils::flatten_schema_arg;
-use crate::CustomTypeSet;
 
 pub fn entry_point_args(entry_point: &Entrypoint, types: &CustomTypeSet) -> Vec<CommandArg> {
     entry_point
